+++ PyPort/PyPort.h
@@ -60,14 +60,14 @@ public:
 	void Event(std::shared_ptr<const EventInfo> event, const std::string& SenderName, SharedStatusCallback_t pStatusCallback) override;
 	void SetTimer(uint32_t id, uint32_t delayms);
 	void RestHandler(const std::string& url, const std::string& content, ResponseCallback_t pResponseCallback);
-	void PublishEventCall(const std::string &EventTypeStr, uint32_t ODCIndex, const std::string &QualityStr, const std::string &PayloadStr);
+	void PublishEventCall(const std::string &EventTypeStr, size_t ODCIndex, const std::string &QualityStr, const std::string &PayloadStr);
 	bool IsOperational() { return PortOperational.load(); };
 
 	// Utility/Testing
 	static bool GetEventTypeFromStringName(const std::string StrEventType, EventType& EventTypeResult);
 	static bool GetControlCodeFromStringName(const std::string StrControlCode, ControlCode& ControlCodeResult);
 	static bool GetConnectStateFromStringName(const std::string StrConnectState, ConnectState& ConnectStateResult);
-	static std::shared_ptr<odc::EventInfo> CreateEventFromStrParams(const std::string& EventTypeStr, uint32_t& ODCIndex, const std::string& QualityStr, const std::string& PayloadStr, const std::string &Name);
+	static std::shared_ptr<odc::EventInfo> CreateEventFromStrParams(const std::string& EventTypeStr, size_t & ODCIndex, const std::string& QualityStr, const std::string& PayloadStr, const std::string &Name);
 	static bool GetQualityFlagsFromStringName(const std::string StrQuality, QualityFlags& QualityResult);
 
 	// Keep track of each PyPort so static methods can get access to the correct PyPort instance
